Add way to repeat the first N arguments.
Instead of just numbers or `t` for `:repeatable`, one can also use
a dotted pair of numbers or of a number and `t`.

This allows the first arguments to be repeated in the body,
if we only want repeat a few arguments while using the
`:ensure` keyword.

For example,

    (setup-define :example-local-hook
      (lambda (hook function)
        `(add-hook ',(setup-get 'hook)
                   (lambda ()
                     (add-hook ',hook ,function nil t))))
      :documentation "Add FUNCTION to HOOK only in buffers of the current mode."
      :debug '(symbolp sexp)
      :ensure '(nil func)
      :repeatable '(1 . 1))

lets us write it as `(:example-local-hook hook f1 f2 f3 f4 ...)`
without needing to repeat `hook` in the local macro.